+++ Services/GaussDBforNoSQL/V3/GaussDBforNoSQLAsyncClient.cs
@@ -2069,9 +2069,9 @@ namespace HuaweiCloud.SDK.GaussDBforNoSQL.V3
         }
         
         /// <summary>
-        /// 重启实例的数据库服务
+        /// 重启实例或节点的数据库服务
         ///
-        /// 重启实例的数据库服务。
+        /// 重启实例或节点的数据库服务。
         /// 
         /// Please refer to HUAWEI cloud API Explorer for details.
         /// </summary>