+++ Services/GaussDBforNoSQL/V3/Model/RestartInstanceRequestBody.cs
@@ -1,0 +1,71 @@
+using System;
+using System.Collections.Generic;
+using System.Text;
+using System.Linq;
+using System.Runtime.Serialization;
+
+using Newtonsoft.Json;
+using Newtonsoft.Json.Converters;
+using HuaweiCloud.SDK.Core;
+
+namespace HuaweiCloud.SDK.GaussDBforNoSQL.V3.Model
+{
+    /// <summary>
+    /// 
+    /// </summary>
+    public class RestartInstanceRequestBody 
+    {
+
+        /// <summary>
+        /// 待重启的节点ID。 仅GeminiDB Redis云原生部署模式集群实例支持传入节点ID重启对应节点。 不传则重启整个实例。
+        /// </summary>
+        [JsonProperty("node_id", NullValueHandling = NullValueHandling.Ignore)]
+        public string NodeId { get; set; }
+
+
+
+        /// <summary>
+        /// Get the string
+        /// </summary>
+        public override string ToString()
+        {
+            var sb = new StringBuilder();
+            sb.Append("class RestartInstanceRequestBody {\n");
+            sb.Append("  nodeId: ").Append(NodeId).Append("\n");
+            sb.Append("}\n");
+            return sb.ToString();
+        }
+
+        /// <summary>
+        /// Returns true if objects are equal
+        /// </summary>
+        public override bool Equals(object input)
+        {
+            return this.Equals(input as RestartInstanceRequestBody);
+        }
+
+        /// <summary>
+        /// Returns true if objects are equal
+        /// </summary>
+        public bool Equals(RestartInstanceRequestBody input)
+        {
+            if (input == null) return false;
+            if (this.NodeId != input.NodeId || (this.NodeId != null && !this.NodeId.Equals(input.NodeId))) return false;
+
+            return true;
+        }
+
+        /// <summary>
+        /// Get hash code
+        /// </summary>
+        public override int GetHashCode()
+        {
+            unchecked // Overflow is fine, just wrap
+            {
+                var hashCode = 41;
+                if (this.NodeId != null) hashCode = hashCode * 59 + this.NodeId.GetHashCode();
+                return hashCode;
+            }
+        }
+    }
+}