+++ Services/GaussDBforNoSQL/V3/Model/ShowIpNumRequirementRequest.cs
@@ -31,7 +31,7 @@ namespace HuaweiCloud.SDK.GaussDBforNoSQL.V3.Model
         public string EngineName { get; set; }
 
         /// <summary>
-        /// 实例类型。没有传入实例ID的时候该字段为必传。 - 取值为“Cluster”，表示GeminiDB Cassandra、GeminiDB Influx、GeminiDB Redis集群实例类型。 - 取值为“ReplicaSet”，表示GeminiDB Mongo副本集实例类型。
+        /// 实例类型。没有传入实例ID的时候该字段为必传。   -  取值为“Cluster”，表示GeminiDB Cassandra、GeminiDB Influx、GeminiDB Redis Proxy经典部署模式集群实例类型。   -  取值为“CloudNativeCluster”，表示GeminiDB Cassandra、GeminiDB Influx、GeminiDB Redis云原生部署模式集群实例类型。   -  取值为“RedisCluster”，表示GeminiDB Redis Cluster经典部署模式集群实例类型。   -  取值为“Replication”，表示GeminiDB Redis经典部署模式主备实例类型。   -  取值为“InfluxdbSingle”，表示GeminiDB Influx经典部署模式单节点实例类型。   -  取值为“ReplicaSet”，表示GeminiDB Mongo副本集实例类型。
         /// </summary>
         [SDKProperty("instance_mode", IsQuery = true)]
         [JsonProperty("instance_mode", NullValueHandling = NullValueHandling.Ignore)]