+++ Services/GaussDBforopenGauss/V3/Model/InstanceEngineDetail.cs
@@ -1,0 +1,80 @@
+using System;
+using System.Collections.Generic;
+using System.Text;
+using System.Linq;
+using System.Runtime.Serialization;
+
+using Newtonsoft.Json;
+using Newtonsoft.Json.Converters;
+using HuaweiCloud.SDK.Core;
+
+namespace HuaweiCloud.SDK.GaussDBforopenGauss.V3.Model
+{
+    /// <summary>
+    /// 
+    /// </summary>
+    public class InstanceEngineDetail 
+    {
+
+        /// <summary>
+        /// 引擎版本号。
+        /// </summary>
+        [JsonProperty("engine_version", NullValueHandling = NullValueHandling.Ignore)]
+        public string EngineVersion { get; set; }
+
+        /// <summary>
+        /// 实例详情。
+        /// </summary>
+        [JsonProperty("instances", NullValueHandling = NullValueHandling.Ignore)]
+        public List<InstanceDetail> Instances { get; set; }
+
+
+
+        /// <summary>
+        /// Get the string
+        /// </summary>
+        public override string ToString()
+        {
+            var sb = new StringBuilder();
+            sb.Append("class InstanceEngineDetail {\n");
+            sb.Append("  engineVersion: ").Append(EngineVersion).Append("\n");
+            sb.Append("  instances: ").Append(Instances).Append("\n");
+            sb.Append("}\n");
+            return sb.ToString();
+        }
+
+        /// <summary>
+        /// Returns true if objects are equal
+        /// </summary>
+        public override bool Equals(object input)
+        {
+            return this.Equals(input as InstanceEngineDetail);
+        }
+
+        /// <summary>
+        /// Returns true if objects are equal
+        /// </summary>
+        public bool Equals(InstanceEngineDetail input)
+        {
+            if (input == null) return false;
+            if (this.EngineVersion != input.EngineVersion || (this.EngineVersion != null && !this.EngineVersion.Equals(input.EngineVersion))) return false;
+            if (this.Instances != input.Instances || (this.Instances != null && input.Instances != null && !this.Instances.SequenceEqual(input.Instances))) return false;
+
+            return true;
+        }
+
+        /// <summary>
+        /// Get hash code
+        /// </summary>
+        public override int GetHashCode()
+        {
+            unchecked // Overflow is fine, just wrap
+            {
+                var hashCode = 41;
+                if (this.EngineVersion != null) hashCode = hashCode * 59 + this.EngineVersion.GetHashCode();
+                if (this.Instances != null) hashCode = hashCode * 59 + this.Instances.GetHashCode();
+                return hashCode;
+            }
+        }
+    }
+}