+++ Services/GaussDBforopenGauss/V3/Model/ListScheduleTaskResponse.cs
@@ -1,0 +1,80 @@
+using System;
+using System.Collections.Generic;
+using System.Text;
+using System.Linq;
+using System.Runtime.Serialization;
+
+using Newtonsoft.Json;
+using Newtonsoft.Json.Converters;
+using HuaweiCloud.SDK.Core;
+
+namespace HuaweiCloud.SDK.GaussDBforopenGauss.V3.Model
+{
+    /// <summary>
+    /// Response Object
+    /// </summary>
+    public class ListScheduleTaskResponse : SdkResponse
+    {
+
+        /// <summary>
+        /// 任务列表。
+        /// </summary>
+        [JsonProperty("tasks", NullValueHandling = NullValueHandling.Ignore)]
+        public List<ScheduleTaskDetail> Tasks { get; set; }
+
+        /// <summary>
+        /// 任务数量。
+        /// </summary>
+        [JsonProperty("total_count", NullValueHandling = NullValueHandling.Ignore)]
+        public int? TotalCount { get; set; }
+
+
+
+        /// <summary>
+        /// Get the string
+        /// </summary>
+        public override string ToString()
+        {
+            var sb = new StringBuilder();
+            sb.Append("class ListScheduleTaskResponse {\n");
+            sb.Append("  tasks: ").Append(Tasks).Append("\n");
+            sb.Append("  totalCount: ").Append(TotalCount).Append("\n");
+            sb.Append("}\n");
+            return sb.ToString();
+        }
+
+        /// <summary>
+        /// Returns true if objects are equal
+        /// </summary>
+        public override bool Equals(object input)
+        {
+            return this.Equals(input as ListScheduleTaskResponse);
+        }
+
+        /// <summary>
+        /// Returns true if objects are equal
+        /// </summary>
+        public bool Equals(ListScheduleTaskResponse input)
+        {
+            if (input == null) return false;
+            if (this.Tasks != input.Tasks || (this.Tasks != null && input.Tasks != null && !this.Tasks.SequenceEqual(input.Tasks))) return false;
+            if (this.TotalCount != input.TotalCount || (this.TotalCount != null && !this.TotalCount.Equals(input.TotalCount))) return false;
+
+            return true;
+        }
+
+        /// <summary>
+        /// Get hash code
+        /// </summary>
+        public override int GetHashCode()
+        {
+            unchecked // Overflow is fine, just wrap
+            {
+                var hashCode = 41;
+                if (this.Tasks != null) hashCode = hashCode * 59 + this.Tasks.GetHashCode();
+                if (this.TotalCount != null) hashCode = hashCode * 59 + this.TotalCount.GetHashCode();
+                return hashCode;
+            }
+        }
+    }
+}