+++ Services/Ocr/V1/Model/BankReceiptRequestBody.cs
@@ -34,6 +34,18 @@ namespace HuaweiCloud.SDK.Ocr.V1.Model
         [JsonProperty("page_num", NullValueHandling = NullValueHandling.Ignore)]
         public int? PageNum { get; set; }
 
+        /// <summary>
+        /// 单朝向模式开关。可选值包括： - true：打开单朝向模式。 - false：关闭单朝向模式。  图片文字方向一致时，打开该开关可提升识别精度；图片文字方向不一致时，关闭该开关可支持多朝向文字识别。未传入该参数时默认为\&quot;true\&quot;，既默认图片中的文字方向为单朝向。 
+        /// </summary>
+        [JsonProperty("single_orientation_mode", NullValueHandling = NullValueHandling.Ignore)]
+        public bool? SingleOrientationMode { get; set; }
+
+        /// <summary>
+        /// 是否打开印章擦除功能。可选值包括： - true：打开印章擦除功能。 - false：关闭印章擦除功能。  开启后，可提升印章遮挡区域的文字识别精度。 
+        /// </summary>
+        [JsonProperty("erase_seal", NullValueHandling = NullValueHandling.Ignore)]
+        public bool? EraseSeal { get; set; }
+
 
 
         /// <summary>
@@ -46,6 +58,8 @@ namespace HuaweiCloud.SDK.Ocr.V1.Model
             sb.Append("  data: ").Append(Data).Append("\n");
             sb.Append("  url: ").Append(Url).Append("\n");
             sb.Append("  pageNum: ").Append(PageNum).Append("\n");
+            sb.Append("  singleOrientationMode: ").Append(SingleOrientationMode).Append("\n");
+            sb.Append("  eraseSeal: ").Append(EraseSeal).Append("\n");
             sb.Append("}\n");
             return sb.ToString();
         }
@@ -67,6 +81,8 @@ namespace HuaweiCloud.SDK.Ocr.V1.Model
             if (this.Data != input.Data || (this.Data != null && !this.Data.Equals(input.Data))) return false;
             if (this.Url != input.Url || (this.Url != null && !this.Url.Equals(input.Url))) return false;
             if (this.PageNum != input.PageNum || (this.PageNum != null && !this.PageNum.Equals(input.PageNum))) return false;
+            if (this.SingleOrientationMode != input.SingleOrientationMode || (this.SingleOrientationMode != null && !this.SingleOrientationMode.Equals(input.SingleOrientationMode))) return false;
+            if (this.EraseSeal != input.EraseSeal || (this.EraseSeal != null && !this.EraseSeal.Equals(input.EraseSeal))) return false;
 
             return true;
         }
@@ -82,6 +98,8 @@ namespace HuaweiCloud.SDK.Ocr.V1.Model
                 if (this.Data != null) hashCode = hashCode * 59 + this.Data.GetHashCode();
                 if (this.Url != null) hashCode = hashCode * 59 + this.Url.GetHashCode();
                 if (this.PageNum != null) hashCode = hashCode * 59 + this.PageNum.GetHashCode();
+                if (this.SingleOrientationMode != null) hashCode = hashCode * 59 + this.SingleOrientationMode.GetHashCode();
+                if (this.EraseSeal != null) hashCode = hashCode * 59 + this.EraseSeal.GetHashCode();
                 return hashCode;
             }
         }